src/jobapplier/jobspy-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jobspy>=1.1.3
pydantic>=2.5.0
pandas>=2.0.0